# backend/openapi.yaml
openapi: 3.0.0
info:
  title: CyberRange Aggregator (Minimal)
  version: "0.1.0"
  description: |-
    Minimal OpenAPI spec for the CyberRange aggregator used by the
    frontend. This is intentionally small — students should extend it
    as they implement the backend.
servers:
  - url: http://localhost:8000
paths:
...
  /api/flags:
    post:
      summary: Submit a flag
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                teamId:
                  type: string
                challengeId:
                  type: string
                flag:
                  type: string
      responses:
        "200":
          description: Flag check response
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  message:
                    type: string
  /api/telemetry:
    post:
      summary: Push telemetry from a Pi agent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: "#/components/schemas/TelemetryEnvelope"
                - $ref: "#/components/schemas/TelemetryPoint"
      responses:
        "200":
          description: accepted
...
components:
  schemas:
...
    TelemetryPoint:
      type: object
      description: Raw telemetry datapoint (agent payload)
      properties:
        nodeId:
          type: string
        host:
          type: string
        ip:
          type: string
        ts:
          type: integer
        cpu:
          type: number
        memFreeMb:
          type: number
        rx:
          type: number
        tx:
          type: number
      additionalProperties: true
    TelemetryEnvelope:
      type: object
      description: Wrapper expected by some aggregator implementations
      properties:
        nodeId:
          type: string
        payload:
          $ref: "#/components/schemas/TelemetryPoint"
...